requirements.txt
line-bot-sdk
flask[async]
pillow
